+++ react-components/src/components/Table/model/i-cell-component.ts
@@ -1,11 +1,13 @@
 import { Column, Row, Table } from '@tanstack/react-table';
 import { IDictionary } from '../model/public-api';
 
-export interface ICellComponent<T = any> {
+export interface ICellComponent<T> {
   row: Row<T>;
   column: Column<T>;
   table: Table<T>;
   dict: IDictionary | null;
   isEdit: boolean;
   cellIndex: number;
+  dateFormat?: string;
+  roundToDecimalPlaces?: number;
 }