+++ react-components/src/components/Table/model/index.ts
@@ -42,6 +42,9 @@ export interface TableProps<T>
   // Обратный вызов изменения значения в ячейке таблицы
   valueChange?: (cell: ICellInstance<T>) => void;
 
-  // Список имен столбцов, которые нужно дополнительно скрыть
-  hiddenColumnNames?: string[];
+  // Форматирование ячеек давт по умолчанию
+  dateFormat?: string;
+
+  // Округлить числа до заданного количества знаков после точки
+  roundToDecimalPlaces?: number;
 }